# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assistant IT",
        "description": "\r\n *              Основная информация:\r\n *              Создатель документации: Егор Лисица\r\n *              Создана специально для: Димы Логунова",
        "version": "1.0"
    },
    "paths": {
        "/api/admin/create/subject": {
            "post": {
                "tags": [
                    "Subject"
                ],
                "summary": "Create Subject",
                "description": "Создает тематику для боковой панели навигации",
                "operationId": "e54b763308e6b37d371db6d9dbdb3d78",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "theme_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Laravel",
                                    "theme_id": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/admin/delete/subject": {
            "delete": {
                "tags": [
                    "Subject"
                ],
                "summary": "Delete Subject",
                "description": "Удаляет тематику для боковой панели навигации",
                "operationId": "a05b91f006958c0c2cc53791b0755637",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Laravel"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/admin/update/subject": {
            "put": {
                "tags": [
                    "Subject"
                ],
                "summary": "Update Subject",
                "description": "Обновляет тематику для боковой панели навигации",
                "operationId": "001aa413e6dddbba801db5f2625a8bc6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "titleUpdate": {
                                                "type": "string"
                                            },
                                            "theme_id": {
                                                "type": "integer"
                                            },
                                            "theme_idUpdate": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Laravel",
                                    "titleUpdate": "Vue",
                                    "theme_id": "1",
                                    "theme_idUpdate": "3"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/admin/create/theme": {
            "post": {
                "tags": [
                    "Theme"
                ],
                "summary": "Create Theme",
                "description": "Создает тему для боковой панели навигации",
                "operationId": "fb08ad6999293fc895b738e3723baeab",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Языки программирования"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "title"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/delete/theme": {
            "delete": {
                "tags": [
                    "Theme"
                ],
                "summary": "Delete Theme",
                "description": "Удаляет тему для боковой панели навигации",
                "operationId": "6eeeefa6d667e8b2ffc278a9f170ce2d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Языки программирования"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/admin/update/theme": {
            "put": {
                "tags": [
                    "Theme"
                ],
                "summary": "Update Theme",
                "description": "Обновляет тему для боковой панели навигации",
                "operationId": "5b8daf08cbac8be9b700f9209316f42e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "titleUpdate": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Языки программирования",
                                    "titleUpdate": "Laravel"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/forum/aside/get": {
            "get": {
                "tags": [
                    "Aside"
                ],
                "summary": "Get Aside",
                "description": "Возвращает ответ с темами и тематиками Aside (Боковая меню слево)",
                "operationId": "9d65c329b49cc544f032dff06c0bcfe7",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "subject": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": "1",
                                    "title": "Фреймворки",
                                    "subjects": {
                                        "id": "7",
                                        "title": "Laravel"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}